statement-store: extract PeersIndex into its own module
